+++ wallet/src/ui/app/components/navigation/index.tsx
@@ -7,16 +7,12 @@ import { NavLink } from 'react-router-dom';
 
 import Icon, { SuiIcons } from '_components/icon';
 import { useAppSelector } from '_hooks';
-import { getNavIsVisible, getNavFilterTag } from '_redux/slices/app';
+import { getNavIsVisible } from '_redux/slices/app';
 
 import st from './Navigation.module.scss';
 
 function makeLinkCls({ isActive }: { isActive: boolean }) {
     return cl(st.link, { [st.active]: isActive });
-}
-
-function activeTagsFilter({ isActive }: { isActive: boolean }) {
-    return cl({ [st.active]: isActive }, st.filter);
 }
 
 export type NavigationProps = {
@@ -25,28 +21,11 @@ export type NavigationProps = {
 
 function Navigation({ className }: NavigationProps) {
     const isVisible = useAppSelector(getNavIsVisible);
-
-    // TODO - use different method for the filters
-    const filterTags = useAppSelector(getNavFilterTag);
     return (
         <nav
             className={cl(st.container, className, { [st.hidden]: !isVisible })}
         >
-            {filterTags.length ? (
-                <div className={st.filterTags}>
-                    {filterTags.map((tag, index) => (
-                        <NavLink
-                            key={tag.link}
-                            to={`./${tag.link}`}
-                            end
-                            className={activeTagsFilter}
-                            title={tag.name}
-                        >
-                            <span className={st.title}>{tag.name}</span>
-                        </NavLink>
-                    ))}
-                </div>
-            ) : null}
+            <div id="sui-apps-filters"></div>
 
             <div className={st.navMenu}>
                 <NavLink to="./tokens" className={makeLinkCls} title="Tokens">
