+++ wallet/src/ui/app/pages/home/apps/ConnectedApps.tsx
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { Content } from '_app/shared/bottom-menu-layout';
-import { ConnectedAppsCard } from '_components/sui-apps';
+import { ConnectedAppsCard, AppFiltersPortal } from '_components/sui-apps';
 
 import st from './AppsPage.module.scss';
 
@@ -11,6 +11,7 @@ function SuiConnectedAppsPage() {
         <div className={st.container}>
             <Content>
                 <section>
+                    <AppFiltersPortal />
                     <ConnectedAppsCard />
                 </section>
             </Content>
