+++ wallet/src/ui/app/pages/home/apps/index.tsx
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { Content } from '_app/shared/bottom-menu-layout';
-import AppsPlayGround from '_components/sui-apps';
+import AppsPlayGround, { AppFiltersPortal } from '_components/sui-apps';
 
 import st from './AppsPage.module.scss';
 
@@ -11,6 +11,7 @@ function AppsPage() {
         <div className={st.container}>
             <Content>
                 <section>
+                    <AppFiltersPortal />
                     <AppsPlayGround />
                 </section>
             </Content>
